Fix overflow in reduce. (#208)
* Fix overflow in reduce.

* Fix doctests as Python 2.7 int is still 64-bit.

+++ sparse/compatibility.py
@@ -14,5 +14,5 @@ else:
     from itertools import izip_longest as zip_longest
     from itertools import izip as zip
 
-    int = long
+    int = int
     range = xrange
